+++ pkg/local_object_storage/blobstor/fstree/fstree.go
@@ -473,9 +473,15 @@ func (t *FSTree) GetRange(addr oid.Address, from uint64, length uint64) ([]byte,
 	}
 
 	payload := obj.Payload()
-	to := from + length
+	pLen := uint64(len(payload))
+	var to uint64
+	if length != 0 {
+		to = from + length
+	} else {
+		to = pLen
+	}
 
-	if pLen := uint64(len(payload)); to < from || pLen < from || pLen < to {
+	if to < from || pLen < from || pLen < to {
 		return nil, logicerr.Wrap(apistatus.ObjectOutOfRange{})
 	}