+++ pkg/services/object/get/assemble.go
@@ -111,8 +111,11 @@ func (exec *execCtx) initFromChild(obj oid.ID) (prev *oid.ID, children []oid.ID)
 	if rng := exec.ctxRange(); rng != nil {
 		seekOff := rng.GetOffset()
 		seekLen := rng.GetLength()
-		seekTo := seekOff + seekLen
 		parSize := par.PayloadSize()
+		if seekLen == 0 {
+			seekLen = parSize
+		}
+		seekTo := seekOff + seekLen
 
 		if seekTo < seekOff || parSize < seekOff || parSize < seekTo {
 			var errOutOfRange apistatus.ObjectOutOfRange