+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Changelog for NeoFS Node
 - New network map support solving the limit of ~320 nodes per network
 - Calculation of VUB for zero hash (#3134)
 - More efficient block header subscription is used now instead of block-based one (#3163)
+- `ObjectService.GetRange(Hash)` ops now handle zero ranges as full payload (#3071)
 
 ### Removed
 - Drop creating new eacl tables with public keys (#3096)